% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarify-package.R
\docType{package}
\name{clarify-package}
\alias{clarify}
\alias{clarify-package}
\title{clarify: Simulation-Based Inference for Regression Models}
\description{
Performs simulation-based inference as an alternative to the delta method for obtaining valid confidence intervals and p-values for regression post-estimation quantities, such as average marginal effects and predictions at representative values. This framework for simulation-based inference is especially useful when the resulting quantity is not normally distributed and the delta method approximation fails. The methodology is described in Greifer, et al. (2025) \doi{10.32614/RJ-2024-015}. 'clarify' is meant to replace some of the functionality of the archived package 'Zelig'; see the vignette "Translating Zelig to clarify" for replicating this functionality.
}
\references{
Greifer, N., Worthington, S., Iacus, S., & King, G. (2025). clarify: Simulation-Based Inference for Regression Models. \emph{The R Journal} 16(2), 154–174. \doi{10.32614/RJ-2024-015}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iqss/clarify}
  \item \url{https://iqss.github.io/clarify/}
  \item Report bugs at \url{https://github.com/iqss/clarify/issues}
}

}
\author{
\strong{Maintainer}: Noah Greifer \email{ngreifer@iq.harvard.edu} (\href{https://orcid.org/0000-0003-3067-7154}{ORCID})

Authors:
\itemize{
  \item Noah Greifer \email{ngreifer@iq.harvard.edu} (\href{https://orcid.org/0000-0003-3067-7154}{ORCID})
  \item Steven Worthington \email{sworthington@iq.harvard.edu} (\href{https://orcid.org/0000-0001-9550-5797}{ORCID})
  \item Stefano Iacus (\href{https://orcid.org/0000-0002-4884-0047}{ORCID})
  \item Gary King \email{king@harvard.edu} (\href{https://orcid.org/0000-0002-5327-7631}{ORCID})
}

}
\keyword{internal}
