% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simbased_est.R
\name{plot.simbased_est}
\alias{plot.simbased_est}
\title{Plot distribution of simulated estimates}
\usage{
\method{plot}{simbased_est}(x, est, ci = TRUE, alpha = 0.05, normal = FALSE)
}
\arguments{
\item{x}{a \code{simbased_est} object; the output of a call to \code{sim_apply()}.}

\item{est}{a vector of the names or indices of the estimates to plot. If unspecified, all estimates will be plotted.}

\item{ci}{\code{logical}; whether to display confidence interval limits for the estimates. Default is \code{TRUE}.}

\item{alpha}{the \eqn{\alpha} used to compute the confidence interval. Default is .05 for 95\% confidence intervals.}

\item{normal}{\code{logical}; whether to compute confidence intervals using a normal approximation (\code{TRUE}) or the simulated sampling distribution (\code{FALSE}; default). See \code{\link[=summary.simbased_est]{summary.simbased_est()}} for details.}
}
\value{
A \code{ggplot} object.
}
\description{
\code{plot()} plots the distribution of the simulated estimates in a density plot.
}
\details{
These plots are produced using \code{\link[ggplot2:geom_density]{ggplot2::geom_density()}}.
}
