% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simbased-package.R
\docType{package}
\name{simbased-package}
\alias{simbased}
\alias{simbased-package}
\title{simbased: Simulation-Based Inference for Regression Models}
\description{
Performs simulation-based inference as an alternative to the delta method for obtaining valid confidence intervals and p-values, which is especially useful when the resulting quantity is not normally distributed and the delta method approximation fails. The framework for simulation-based inference is described in King, Tomz, and Wittenberg (2000) \doi{10.2307/2669316}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ngreifer/simbased}
  \item Report bugs at \url{https://github.com/ngreifer/simbased/issues}
}

}
\author{
\strong{Maintainer}: Noah Greifer \email{ngreifer@iq.harvard.edu} (\href{https://orcid.org/0000-0003-3067-7154}{ORCID}) [contributor]

}
\keyword{internal}
