% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Simulate model coefficients}
\usage{
sim(fit, n = 1000, vcov = stats::vcov, coefs = stats::coef, dist = NULL)
}
\arguments{
\item{fit}{a model fit, such as the output of a call to `[lm()]` or `[glm()]`. Can be left unspecified if `coefs` and `vcov` are not functions.}

\item{n}{the number of simulations to run; default is 1000. More is always better but resulting calculations will take longer.}

\item{vcov}{either a square covariance matrix of the parameter estimates or a function to use to extract it from `fit`. By default, uses `[stats::vcov()]`.}

\item{coefs}{either a vector of coefficient estimates or a function to use to extract it from `fit`. By default, uses `[stats::coef()]`. Using anything but the default changes some options in `sim_apply()`; see `[sim_apply()]`'s Details section for more information.}

\item{dist}{a string containing the name of the multivariate distribution to use to draw simulated coefficients. Should be one of `"normal"` (multivariate normal distribution) or `"t_{#}"` (multivariate t distribution), where `{#}` corresponds to the desired degrees of freedom (e.g., `"t_100"`). If `NULL`, `sim()` will attempt to figure out the right distribution to use; see Details.}
}
\value{
A `simbased_sim` object, which has the following components:
 \item{coefs}{a matrix containing the simulated coefficients with a column for each coefficient and a row for each simulation}
 \item{fit}{the original model fit supplied to}
The `"dist"` attribute contains `"normal"` if the coefficients were sampled from a multivariate normal distribution and `"t"` if sampled from a mutivariate t distribution. The `"simbased_hash"` attribute contains a unique hash generated by [rlang::hash()].
}
\description{
`sim()` simulates model parameters from a multivariate normal or t distribution that are then used by `[sim_apply()]` to calculate quantities of interest.
}
\details{
When `dist` is `NULL`, `sim()` samples from a multivariate normal or t distribution depending on whether `df.residual(fit)` is `NULL` or not. If `NULL` or `fit` is unspecified, a multivariate normal distribution is used. Otherwise (primarily for linear models fit with `lm()`), a multivariate t distribution is used, with `df.residual(fit)` used as the degrees of freedom.

When a multivariate normal is used, it is sampled from with means equal to the estimated coefficients and the parameter covariance matrix as the covariance matrix using `[mvnfast::rmvn()]`. When a multivariate t distribution is used, it is sampled from with means equal to the estimated coefficients and scaling matrix equal to `cov*(df - 2)/df`, where `cov` is the parameter covariance matrix and `df` is the residual degrees of freedom for the model, using `[mvnfast::rmvt()]`.
}
\seealso{
`[sim_apply()]` for applying a function to each set of simulated coefficients.
}
