% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Simulate Model Coefficients}
\usage{
sim(fit, n = 1000, vcov = stats::vcov)
}
\arguments{
\item{fit}{a model fit, such as the output of a call to `lm()` or `glm()`}

\item{n}{the number of simulations to run; default is 1000. More is better but resulting calculations will take longer.}

\item{vcov}{optional; a function to use to extract the asymptotic covariance matrix of the parameter estimates. By default, uses `stats::vcov()`.}
}
\value{
A `simbased_sim` object, which has the following components:
}
\description{
Simulate Model Coefficients
}
