% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_apply.R
\name{sim_apply}
\alias{sim_apply}
\title{Apply a Function To Simulated Parameter Values}
\usage{
sim_apply(sim, FUN, ...)
}
\arguments{
\item{...}{}
}
\description{
Apply a Function To Simulated Parameter Values
}
