% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.simbased_est.R
\name{summary.simbased_est}
\alias{summary.simbased_est}
\title{Title}
\usage{
\method{summary}{simbased_est}(
  object,
  alpha = 0.05,
  normal = FALSE,
  test = FALSE,
  null = 0,
  transform = NULL,
  ...
)
}
\arguments{
\item{object}{a `simbased_est` object; the output of a call to `sim_apply()`.}

\item{alpha}{the $\alpha$ used to compute the confidence interval. Default is .05 for 95% confidence intervals.}

\item{normal}{`logical`; whether to compute p-values and confidence intervals using a normal approximation (`TRUE`) or the simulated sampling distribution (`FALSE`; default).}

\item{null}{the value of the parameter under the null hypothesis for the p-value calculations.}

\item{...}{ignored.}
}
\description{
Title
}
